+++ CHANGES.md
@@ -1,6 +1,7 @@
 HEAD
 -----------
 
+- When nesting, make parent accessible to child.
 - Craft#attributes returns the attributes.
 - Craft.stub stubs a static or dynamic value that doesn't need to be parsed in
   the document.